[chore] rename release to RC where applicable
